requirements.txt
google-generativeai
langchain
PyPDF2
faiss-cpu
langchain_google_genai
pypdf
langchain-community
ipykernel
Flask
Flask-Cors
python-dotenv